server/requirements.txt
adium-theme-ubuntu==0.3.4
Django==1.10.2
django-cors-headers==1.2.2
djangorestframework==3.4.7
Pillow==3.4.1
unity-lens-photos==1.0
virtualenv==15.0.3
whitenoise==3.2.2
